change unit test to use mock logger, adjust log level to warning
