+++ tests/unit/utils/test_nova_compute.py
@@ -11,8 +11,7 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-import logging
-from unittest.mock import AsyncMock, MagicMock, patch
+from unittest.mock import AsyncMock, MagicMock, call, patch
 
 import pytest
 from juju.action import Action
@@ -64,21 +63,30 @@ async def test_get_instance_count_invalid_result(model, result_key, value):
     ],
 )
 @pytest.mark.asyncio
+@patch("cou.utils.nova_compute.logger")
 @patch("cou.utils.nova_compute.get_instance_count")
 async def test_get_empty_hypervisors(
-    mock_instance_count, hypervisors_count, expected_result, model, caplog
+    mock_instance_count, mock_logger, hypervisors_count, expected_result, model
 ):
     mock_instance_count.side_effect = [count for _, count in hypervisors_count]
-    non_empty_log, selected_log = construct_expected_logs(hypervisors_count)
-
-    with caplog.at_level(logging.INFO):
-        result = await nova_compute.get_empty_hypervisors(
-            [_mock_nova_unit(nova_unit) for nova_unit, _ in hypervisors_count], model
-        )
+    result = await nova_compute.get_empty_hypervisors(
+        [_mock_nova_unit(nova_unit) for nova_unit, _ in hypervisors_count], model
+    )
 
     assert {machine.machine_id for machine in result} == expected_result
-    assert non_empty_log in caplog.record_tuples
-    assert selected_log in caplog.record_tuples
+
+    selected_hypervisors = ", ".join(
+        [f"nova-compute/{i}" for i, count in hypervisors_count if count == 0]
+    )
+    non_empty_hypervisors = ", ".join(
+        [f"nova-compute/{i}" for i, count in hypervisors_count if count != 0]
+    )
+
+    mock_logger.info.assert_has_calls([call("Selected hypervisors: %s", selected_hypervisors)])
+    if non_empty_hypervisors:
+        mock_logger.warning.assert_has_calls(
+            [call("Found non-empty hypervisors: %s", non_empty_hypervisors)]
+        )
 
 
 @pytest.mark.parametrize("instance_count", [1, 10, 50])
@@ -112,24 +120,3 @@ def _mock_nova_unit(nova_unit):
     mock_nova_unit.machine = nova_machine
 
     return mock_nova_unit
-
-
-def construct_expected_logs(hypervisors_count):
-    selected_hypervisors = [f"nova-compute/{i}" for i, count in hypervisors_count if count == 0]
-    non_empty_hypervisors = [f"nova-compute/{i}" for i, count in hypervisors_count if count != 0]
-
-    selected_hypervisors_str = ", ".join(selected_hypervisors)
-    non_empty_hypervisors_str = ", ".join(non_empty_hypervisors)
-
-    selected_log = (
-        "cou.utils.nova_compute",
-        logging.INFO,
-        f"Selected hypervisors: {selected_hypervisors_str}",
-    )
-    non_empty_log = (
-        "cou.utils.nova_compute",
-        logging.INFO,
-        f"Found non-empty hypervisors: {non_empty_hypervisors_str}",
-    )
-
-    return non_empty_log, selected_log
